# src/PaperlessREST/Controllers/ConfigApi.cs
/*
 * Paperless Rest Server
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using PaperlessREST.Attributes;
using PaperlessREST.Entities;

namespace PaperlessREST.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ConfigApiController : ControllerBase
    { 
        /// <summary>
        /// 
        /// </summary>
        /// <param name="createSavedViewsRequest"></param>
        /// <response code="200">Success</response>
        [HttpPost]
        [Route("/api/saved_views")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("CreateSavedViews")]
        public virtual IActionResult CreateSavedViews([FromBody]CreateSavedViewsRequest createSavedViewsRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="createStoragePathRequest"></param>
        /// <response code="200">Success</response>
        [HttpPost]
        [Route("/api/storage_paths")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("CreateStoragePath")]
        [SwaggerResponse(statusCode: 200, type: typeof(CreateStoragePath200Response), description: "Success")]
        public virtual IActionResult CreateStoragePath([FromBody]CreateStoragePathRequest createStoragePathRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(CreateStoragePath200Response));
            string exampleJson = null;
            exampleJson = "{\r\n  \"owner\" : 1,\r\n  \"path\" : \"path\",\r\n  \"matching_algorithm\" : 6,\r\n  \"user_can_change\" : true,\r\n  \"is_insensitive\" : true,\r\n  \"name\" : \"name\",\r\n  \"match\" : \"match\",\r\n  \"id\" : 0,\r\n  \"slug\" : \"slug\"\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<CreateStoragePath200Response>(exampleJson)
            : default(CreateStoragePath200Response);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="createUISettingsRequest"></param>
        /// <response code="200">Success</response>
        [HttpPost]
        [Route("/api/ui_settings")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("CreateUISettings")]
        [SwaggerResponse(statusCode: 200, type: typeof(CreateUISettings200Response), description: "Success")]
        public virtual IActionResult CreateUISettings([FromBody]CreateUISettingsRequest createUISettingsRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(CreateUISettings200Response));
            string exampleJson = null;
            exampleJson = "{\r\n  \"success\" : true\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<CreateUISettings200Response>(exampleJson)
            : default(CreateUISettings200Response);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="id"></param>
        /// <response code="204">Success</response>
        [HttpDelete]
        [Route("/api/storage_paths/{id}")]
        [ValidateModelState]
        [SwaggerOperation("DeleteStoragePath")]
        public virtual IActionResult DeleteStoragePath([FromRoute (Name = "id")][Required]int id)
        {

            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);

            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/ws/status")]
        [ValidateModelState]
        [SwaggerOperation("Get")]
        public virtual IActionResult Get()
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            throw new NotImplementedException();
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="id"></param>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/api/logs/{id}")]
        [ValidateModelState]
        [SwaggerOperation("GetLog")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<string>), description: "Success")]
        public virtual IActionResult GetLog([FromRoute (Name = "id")][Required]string id)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<string>));
            string exampleJson = null;
            exampleJson = "[ \"\", \"\" ]";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<string>>(exampleJson)
            : default(List<string>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/api/logs")]
        [ValidateModelState]
        [SwaggerOperation("GetLogs")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<string>), description: "Success")]
        public virtual IActionResult GetLogs()
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<string>));
            string exampleJson = null;
            exampleJson = "[ \"\", \"\" ]";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<string>>(exampleJson)
            : default(List<string>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="page"></param>
        /// <param name="pageSize"></param>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/api/saved_views")]
        [ValidateModelState]
        [SwaggerOperation("GetSavedViews")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetSavedViews200Response), description: "Success")]
        public virtual IActionResult GetSavedViews([FromQuery (Name = "page")]int? page, [FromQuery (Name = "page_size")]int? pageSize)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(GetSavedViews200Response));
            string exampleJson = null;
            exampleJson = "{\r\n  \"next\" : 6,\r\n  \"all\" : [ 5, 5 ],\r\n  \"previous\" : 1,\r\n  \"count\" : 0,\r\n  \"results\" : [ {\r\n    \"owner\" : {\r\n      \"is_superuser\" : true,\r\n      \"is_active\" : true,\r\n      \"user_permissions\" : [ 9, 9 ],\r\n      \"is_staff\" : true,\r\n      \"last_login\" : \"last_login\",\r\n      \"last_name\" : \"last_name\",\r\n      \"groups\" : [ \"\", \"\" ],\r\n      \"password\" : \"password\",\r\n      \"id\" : 7,\r\n      \"date_joined\" : \"date_joined\",\r\n      \"first_name\" : \"first_name\",\r\n      \"email\" : \"email\",\r\n      \"username\" : \"username\"\r\n    },\r\n    \"user_can_change\" : true,\r\n    \"sort_field\" : \"sort_field\",\r\n    \"show_on_dashboard\" : true,\r\n    \"name\" : \"name\",\r\n    \"show_in_sidebar\" : true,\r\n    \"filter_rules\" : [ {\r\n      \"rule_type\" : 2,\r\n      \"value\" : \"value\"\r\n    }, {\r\n      \"rule_type\" : 2,\r\n      \"value\" : \"value\"\r\n    } ],\r\n    \"sort_reverse\" : true,\r\n    \"id\" : 5\r\n  }, {\r\n    \"owner\" : {\r\n      \"is_superuser\" : true,\r\n      \"is_active\" : true,\r\n      \"user_permissions\" : [ 9, 9 ],\r\n      \"is_staff\" : true,\r\n      \"last_login\" : \"last_login\",\r\n      \"last_name\" : \"last_name\",\r\n      \"groups\" : [ \"\", \"\" ],\r\n      \"password\" : \"password\",\r\n      \"id\" : 7,\r\n      \"date_joined\" : \"date_joined\",\r\n      \"first_name\" : \"first_name\",\r\n      \"email\" : \"email\",\r\n      \"username\" : \"username\"\r\n    },\r\n    \"user_can_change\" : true,\r\n    \"sort_field\" : \"sort_field\",\r\n    \"show_on_dashboard\" : true,\r\n    \"name\" : \"name\",\r\n    \"show_in_sidebar\" : true,\r\n    \"filter_rules\" : [ {\r\n      \"rule_type\" : 2,\r\n      \"value\" : \"value\"\r\n    }, {\r\n      \"rule_type\" : 2,\r\n      \"value\" : \"value\"\r\n    } ],\r\n    \"sort_reverse\" : true,\r\n    \"id\" : 5\r\n  } ]\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<GetSavedViews200Response>(exampleJson)
            : default(GetSavedViews200Response);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="page"></param>
        /// <param name="fullPerms"></param>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/api/storage_paths")]
        [ValidateModelState]
        [SwaggerOperation("GetStoragePaths")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetStoragePaths200Response), description: "Success")]
        public virtual IActionResult GetStoragePaths([FromQuery (Name = "page")]int? page, [FromQuery (Name = "full_perms")]bool? fullPerms)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(GetStoragePaths200Response));
            string exampleJson = null;
            exampleJson = "{\r\n  \"next\" : true,\r\n  \"all\" : [ 6, 6 ],\r\n  \"previous\" : true,\r\n  \"count\" : 0,\r\n  \"results\" : [ {\r\n    \"owner\" : 2,\r\n    \"path\" : \"path\",\r\n    \"matching_algorithm\" : 5,\r\n    \"document_count\" : 5,\r\n    \"is_insensitive\" : true,\r\n    \"permissions\" : {\r\n      \"view\" : {\r\n        \"groups\" : [ \"\", \"\" ],\r\n        \"users\" : [ 7, 7 ]\r\n      },\r\n      \"change\" : {\r\n        \"groups\" : [ \"\", \"\" ],\r\n        \"users\" : [ 7, 7 ]\r\n      }\r\n    },\r\n    \"name\" : \"name\",\r\n    \"match\" : \"match\",\r\n    \"id\" : 1,\r\n    \"slug\" : \"slug\"\r\n  }, {\r\n    \"owner\" : 2,\r\n    \"path\" : \"path\",\r\n    \"matching_algorithm\" : 5,\r\n    \"document_count\" : 5,\r\n    \"is_insensitive\" : true,\r\n    \"permissions\" : {\r\n      \"view\" : {\r\n        \"groups\" : [ \"\", \"\" ],\r\n        \"users\" : [ 7, 7 ]\r\n      },\r\n      \"change\" : {\r\n        \"groups\" : [ \"\", \"\" ],\r\n        \"users\" : [ 7, 7 ]\r\n      }\r\n    },\r\n    \"name\" : \"name\",\r\n    \"match\" : \"match\",\r\n    \"id\" : 1,\r\n    \"slug\" : \"slug\"\r\n  } ]\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<GetStoragePaths200Response>(exampleJson)
            : default(GetStoragePaths200Response);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/api/ui_settings")]
        [ValidateModelState]
        [SwaggerOperation("GetUISettings")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetUISettings200Response), description: "Success")]
        public virtual IActionResult GetUISettings()
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(GetUISettings200Response));
            string exampleJson = null;
            exampleJson = "{\r\n  \"settings\" : {\r\n    \"update_checking\" : {\r\n      \"backend_setting\" : \"backend_setting\"\r\n    }\r\n  },\r\n  \"permissions\" : [ \"permissions\", \"permissions\" ],\r\n  \"display_name\" : \"display_name\",\r\n  \"user\" : {\r\n    \"is_superuser\" : true,\r\n    \"groups\" : [ \"\", \"\" ],\r\n    \"id\" : 0,\r\n    \"username\" : \"username\"\r\n  }\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<GetUISettings200Response>(exampleJson)
            : default(GetUISettings200Response);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// 
        /// </summary>
        /// <param name="id"></param>
        /// <param name="updateStoragePathRequest"></param>
        /// <response code="200">Success</response>
        [HttpPut]
        [Route("/api/storage_paths/{id}")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("UpdateStoragePath")]
        [SwaggerResponse(statusCode: 200, type: typeof(UpdateStoragePath200Response), description: "Success")]
        public virtual IActionResult UpdateStoragePath([FromRoute (Name = "id")][Required]int id, [FromBody]UpdateStoragePathRequest updateStoragePathRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(UpdateStoragePath200Response));
            string exampleJson = null;
            exampleJson = "{\r\n  \"owner\" : 5,\r\n  \"path\" : \"path\",\r\n  \"matching_algorithm\" : 6,\r\n  \"user_can_change\" : true,\r\n  \"document_count\" : 1,\r\n  \"is_insensitive\" : true,\r\n  \"name\" : \"name\",\r\n  \"match\" : \"match\",\r\n  \"id\" : 0,\r\n  \"slug\" : \"slug\"\r\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<UpdateStoragePath200Response>(exampleJson)
            : default(UpdateStoragePath200Response);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}